Version 4.0 Exception Handling

+++ src/main/java/com/nor/flightManagementSystem/controller/FlightController.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,9 +16,12 @@ import com.nor.flightManagementSystem.bean.Flight;
 import com.nor.flightManagementSystem.bean.Route;
 import com.nor.flightManagementSystem.dao.FlightDao;
 import com.nor.flightManagementSystem.dao.RouteDao;
+import com.nor.flightManagementSystem.exception.DatabaseException;
+import com.nor.flightManagementSystem.exception.DuplicateFlightNumberException;
+import com.nor.flightManagementSystem.exception.FlightNotFoundException;
 import com.nor.flightManagementSystem.service.FlightService;
 
-
+@ControllerAdvice
 @Controller
 public class FlightController {    
     @Autowired
@@ -38,43 +43,64 @@ public class FlightController {
         return mv;
     }
 
-    @PostMapping("/addFlight")
+	@PostMapping("/addFlight")
     public ModelAndView saveFlight(@ModelAttribute("flightRecord") Flight flight, String returnDeparture, String returnArrival) {
-    	Flight returnFlight = flightService.createReturnFlight(flight, returnDeparture, returnArrival);
+    	try {
+    		Long flightNumber = flight.getFlightNumber();
+            if (flightDao.viewFlight(flightNumber) != null) {
+                throw new DuplicateFlightNumberException("Flight with flightNumber " + flightNumber + " already exists.");
+    	} 
+        Flight returnFlight = flightService.createReturnFlight(flight, returnDeparture, returnArrival);
     	flightDao.addFlight(flight);
     	flightDao.addFlight(returnFlight);
         return new ModelAndView("index");
+    	} catch (DuplicateFlightNumberException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new DatabaseException("Error saving flight to the database", e);
+        }
     }
     
     @GetMapping("/viewFlights")
     public ModelAndView showAllFlights() {
+    	try {
         List<Flight> flights = flightDao.showAllFlights();
         ModelAndView mv = new ModelAndView("viewFlightPage");
         mv.addObject("flights", flights);
         return mv;
+    	} catch (Exception e) {
+            throw new DatabaseException("Error retrieving flights from the database", e);
+        }
     }
         
-    
-//  @ExceptionHandler(value = RouteException.class)
-//  public ModelAndView handlingRouteException(RouteException exception) {
-//  	ModelAndView mv = new ModelAndView("routeErrorPage");
-//  	mv.addObject("error", "Arrival And Destinationn Airports can't be same");
-//  	return mv;
-//  }
 
     
     @GetMapping("/modifyFlight")
     public ModelAndView deleteFlight() {
+    	try {
         List<Flight> flights = flightDao.showAllFlights();
         ModelAndView mv = new ModelAndView("modifyFlight");
         mv.addObject("flights", flights);
         return mv;
+    	} catch (Exception e) {
+            throw new DatabaseException("Error retrieving flights from the database", e);
+        }
     }
     
     @PostMapping("/deleteFlight")
     public ModelAndView deleteAirport(@RequestParam("flightNumber") Long flightNumber) {
-    	flightDao.deleteFlightByFlightNumber(flightNumber);
-    	return new ModelAndView("index"); 
+    	try {
+            Flight flight = flightDao.viewFlight(flightNumber);
+            if (flight == null) {
+                throw new FlightNotFoundException("flight with Number " + flightNumber + " not found.");
+            }
+	    	flightDao.deleteFlightByFlightNumber(flightNumber);
+	    	return new ModelAndView("index"); 
+    	} catch (FlightNotFoundException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new DatabaseException("Error deleting flight from the database", e);
+        }
     }
     
     @PostMapping("/updateFlight")
@@ -84,9 +110,45 @@ public class FlightController {
                                @RequestParam("departure") String departure,
                                @RequestParam("routeId") Long routeId,
                                @RequestParam("seatCapacity") Integer seatCapacity) {
-        Flight flight = new Flight(flightNumber, flightName, routeId, seatCapacity, departure, arrival);
-        flightDao.updateFlight(flight);
-        return new ModelAndView("index");
+    	try {
+            Flight flight = flightDao.viewFlight(flightNumber);
+            if (flight == null) {
+                throw new FlightNotFoundException("flight with Number " + flightNumber + " not found.");
+            }
+	        Flight newFlight = new Flight(flightNumber, flightName, routeId, seatCapacity, departure, arrival);
+	        flightDao.updateFlight(newFlight);
+	        return new ModelAndView("index");
+    	} catch (FlightNotFoundException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new DatabaseException("Error updating flight in the database", e);
+        }
     }
 
+    @ExceptionHandler(FlightNotFoundException.class)
+    public ModelAndView handleFlightNotFoundException(FlightNotFoundException e) {
+        ModelAndView mv = new ModelAndView("errorPage");
+        mv.addObject("error", e.getMessage());
+        return mv;
+    } 
+    @ExceptionHandler(DuplicateFlightNumberException.class)
+    public ModelAndView handleDuplicateFlightNumberException(DuplicateFlightNumberException e) {
+        ModelAndView mv = new ModelAndView("errorPage");
+        mv.addObject("error", e.getMessage());
+        return mv;
+    }
+
+    @ExceptionHandler(DatabaseException.class)
+    public ModelAndView handleDatabaseException(DatabaseException e) {
+        ModelAndView mv = new ModelAndView("errorPage");
+        mv.addObject("error", e.getMessage());
+        return mv;
+    }
+    
+    @ExceptionHandler(Exception.class)
+    public ModelAndView handleGeneralException(Exception e) {
+        ModelAndView mv = new ModelAndView("errorPage");
+        mv.addObject("error", "An unexpected error occurred. Please try again later.");
+        return mv;
+    }
 }
